Stop publishing insecure disabled-component APIs via zap imports and release artifacts.
Move normalizeURL into common, exclude comment/dm/live-chat from dts, and omit sourcemap sources/maps from the npm package so private-key helpers cannot leak through the release graph.
