+++ src/nostr-comment/utils.ts
@@ -19,32 +19,6 @@ export interface ParsedComment {
 }
 
 /**
- * Normalize URL for consistent comment identification
- */
-export function normalizeURL(raw: string): string {
-    try {
-        const url = new URL(raw);
-        return (
-            url.origin
-                .replace('://m.', '://') // remove known 'mobile' subdomains
-                .replace('://mobile.', '://')
-                .replace('http://', 'https://') // default everything to https
-                .replace(
-                    /:\d+/,
-                    // remove 443 and 80 ports
-                    port => (port === ':443' || port === ':80' ? '' : port)
-                ) +
-            url.pathname
-                .replace(/\/+/g, '/') // remove duplicated slashes in the middle of the path
-                .replace(/\/*$/, '') // remove slashes from the end of path
-        );
-    } catch (error) {
-        console.error('Invalid URL:', raw);
-        return raw;
-    }
-}
-
-/**
  * Parse Nostr event into comment structure
  */
 export function parseCommentEvent(event: NDKEvent): ParsedComment | null {
